make client running as an actor
